+++ src/main/java/io/github/depromeet/knockknockbackend/domain/asset/presentation/dto/response/BackgroundListResponse.java
@@ -1,4 +1,4 @@
-package io.github.depromeet.knockknockbackend.domain.group.presentation.dto.response;
+package io.github.depromeet.knockknockbackend.domain.asset.presentation.dto.response;
 
 import java.util.List;
 import lombok.AllArgsConstructor;
