+++ src/main/java/io/github/depromeet/knockknockbackend/domain/asset/presentation/dto/response/ThumbnailImageDto.java
@@ -1,4 +1,4 @@
-package io.github.depromeet.knockknockbackend.domain.group.presentation.dto.response;
+package io.github.depromeet.knockknockbackend.domain.asset.presentation.dto.response;
 
 import io.github.depromeet.knockknockbackend.domain.group.domain.Thumbnail;
 import lombok.Getter;
